% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalSelTools-package.R
\docType{package}
\name{BalSelTools-package}
\alias{BalSelTools}
\alias{BalSelTools-package}
\title{BalSelTools: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Bárbara Bitarello \email{barbarabitarello@gmail.com} (\href{https://orcid.org/0000-0001-7676-9367}{ORCID})

}
\keyword{internal}
