% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_geno.R
\name{.split_geno}
\alias{.split_geno}
\title{Split a genotype column from a vcf file}
\usage{
.split_geno(x, split = "|")
}
\arguments{
\item{x}{A string from a GT column.}

\item{split}{Pattern to use for splitting.}
}
\description{
Split a genotype column from a vcf file
}
\examples{
split_geno(x = c("0|0","0|1","1|1"), split="|")
}
