% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncd1.R
\name{ncd1}
\alias{ncd1}
\title{Calculate non-central statistic (NCD1)}
\usage{
ncd1(x = x, tf = 0.5, w = 3000, ncores = 2, minIS = 2)
}
\arguments{
\item{x}{A data.table object}

\item{tf}{Target frequency}

\item{w}{Window size in bp. Default is 1000}

\item{ncores}{Number of cores. Increasing this can speed things up for you.
Default is 2.}

\item{minIS}{Minimum number of informative sites. Default is 2. Windows with
less informative sites than this threshold are discarded.}
}
\value{
A data.table object
}
\description{
Calculate non-central statistic (NCD1)
}
\examples{
ncd1(x=ncd1_input, tf=0.5, w=3000, ncores=2, minIS=8)
}
