% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncd1copy.R
\name{ncd1.test}
\alias{ncd1.test}
\title{Calculate non-central statistic (NCD1)}
\usage{
ncd1.test(x = x, tf = 0.5, w = 3000, ncores = 2, minIS = 2, by = "POS")
}
\arguments{
\item{x}{A data.table object with
columns: CHR, POS, REF, ALT, tx_1 (number of alternate allele copies), tn_1 (total number of alleles).}

\item{tf}{Target frequency.}

\item{w}{Window size in bp. Default is 1000.}

\item{ncores}{Number of cores. Increasing this can speed things up for you.
Default is 2.}

\item{minIS}{Minimum number of informative sites. Default is 2. Windows with
less informative sites than this threshold are discarded.}

\item{by}{Define how to scan the genome. "POS" (default) defined sliding windows based on w. "IS" defined windows around each informative site.}
}
\value{
A data.table object with columns:  TO DO
}
\description{
Calculate non-central statistic (NCD1)
}
\examples{
ncd1(x=ncd1_input, tf=0.5, w=3000, ncores=2, minIS=8)
}
