% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncd2_input.R
\docType{data}
\name{ncd2_input}
\alias{ncd2_input}
\title{Input data to run ncd2 example}
\format{
A data.table with 838 rows and 8 columns
\describe{
\item{CHR}{Chromsome, single value - 1}
\item{POS}{Position along the chromosome - first is 92, last is 29992}
\item{REF}{Reference allele - A, C, G, T}
\item{ALT}{Alternate allele - A, C, G, T}
\item{tx_1}{Number of alternate allele copies in the ingroup - min=0, max=216}
\item{tn_1}{Total number of alleles in the ingroup, single value - 216 }
\item{tx_2}{Number of alternate allele copies in the outgroup - min=0, max=4}
\item{tn_2}{Total number of alleles in the outgroup, single value - 216 }
}
}
\source{
created from a VCF (Variant Call Format) input file generated with the SLiM simulator by reading it in with parse_vcf(type="ncd2")
}
\usage{
ncd2_input
}
\description{
An R object in data.table format
}
\keyword{datasets}
