% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outfile_path}
\alias{outfile_path}
\title{Create output file with timestamp}
\usage{
outfile_path(infile = "example.vcf")
}
\arguments{
\item{infile}{Name of input file to modify}
}
\description{
Create output file with timestamp
}
\examples{
outfile_path("inst/example.vcf")
}
