% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vcf.R
\name{parse_vcf}
\alias{parse_vcf}
\title{Parse VCF files generated by SLiM v3}
\usage{
parse_vcf(infile = NULL, outfile = NULL, n0 = NULL, n1 = NULL, type = NULL)
}
\arguments{
\item{infile}{The path and name of a vcf file - single chromosome}

\item{outfile}{The path and name for the outfile. If not provided,
this will be  a timestamp + infile in the current directory.}

\item{n0}{Number of individuals in pop0}

\item{n1}{Number of individuals in pop1}

\item{type}{Which input format will be generated. Options: ncd1, ncd2.}
}
\value{
Returns a data table object.
}
\description{
Parse VCF files generated by SLiM v3
}
\examples{
parse_vcf(infile=system.file(package="balselr", "example.vcf"), n0=108, type="ncd1")
parse_vcf(infile=system.file(package="balselr", "example.vcf"), n0=108, n1=2, type="ncd2")
}
