% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vcf.R
\name{parse_vcf_slim}
\alias{parse_vcf_slim}
\title{Parse VCF files generated by SLiM v3}
\usage{
parse_vcf_slim(
  infile = "*.vcf",
  outfile = NA,
  index.col = NULL,
  nind = c(10, 10, 1),
  type = "ncd2",
  fold = NA,
  intern = T,
  verbose = T
)
}
\arguments{
\item{verbose}{}
}
\description{
Parse VCF files generated by SLiM v3
}
