% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_ncd2.r
\name{vcf_ncd2}
\alias{vcf_ncd2}
\title{Make input file for NCD2 from VCF}
\usage{
vcf_ncd2(
  x = inp,
  outfile = outfile,
  nind = nind,
  index.col = index.col,
  fold = fold,
  verbose = verbose
)
}
\arguments{
\item{x}{A data.table object generated from reading a VCF file with read_vcf.}

\item{outfile}{The path and name for the outfile. If not provided,
this will be  a timestamp file in the current directory.}

\item{nind}{A vector containing the number of diploid individuals from each
population to be used. The length of the vector should be equal to the number
of populations being sampled.}

\item{index.col}{First genotype column in VCF file.}

\item{fold}{Logical. If TRUE, the output will have alternate allele counts.
If FALSE, derived allele counts will be used. Default is TRUE.}

\item{verbose}{Logical. If TRUE, progress reports will be printed as the
function runs.}
}
\description{
Make input file for NCD2 from VCF
}
\examples{
inp = read_vcf("inst/example.vcf")
vcf_ncd2(x=inp, outfile=outfile_path("inst/example.vcf"),nind=c(108,10),
index.col=10, fold=T, verbose=T)

}
