% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_obj.R
\docType{data}
\name{vcf_obj}
\alias{vcf_obj}
\title{An example VCF (variant call format) read into R}
\format{
A data.table with 838 rows and 128 columns
}
\source{
created from a VCF (Variant Call Format) input file generated with the SLiM simulator by reading it in with read_vcf().
}
\usage{
vcf_obj
}
\description{
An example VCF (variant call format) read into R
}
\keyword{datasets}
