% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balselr-package.R
\docType{package}
\name{balselr-package}
\alias{balselr}
\alias{balselr-package}
\title{balselr: Balancing Selection in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Run NCD statistics on population-level genomic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bitarellolab.github.io/balselr/}
  \item Report bugs at \url{https://github.com/bitarellolab/balselr/issues}
}

}
\author{
\strong{Maintainer}: Bárbara Bitarello \email{barbarabitarello@gmail.com} (\href{https://orcid.org/0000-0001-7676-9367}{ORCID})

Authors:
\itemize{
  \item Daphne Hansell [contributor]
}

Other contributors:
\itemize{
  \item Débora YC Brandt [contributor]
}

}
\keyword{internal}
