% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalSelTools-package.R
\docType{package}
\name{balselr-package}
\alias{balselr}
\alias{balselr-package}
\title{balselr: Balancing Selection in R}
\description{
Produce input files for different balancing selection programs/tests and run the tests.
}
\author{
\strong{Maintainer}: Bárbara Bitarello \email{barbarabitarello@gmail.com} (\href{https://orcid.org/0000-0001-7676-9367}{ORCID})

}
\keyword{internal}
