% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalSelTools-package.R
\docType{package}
\name{BalSelTools-package}
\alias{BalSelTools}
\alias{BalSelTools-package}
\title{BalSelTools: A set of functions to perform genome-wide scans and tests for balancing selection in genomic data}
\description{
Easy-to-use, fast, integrated neutrality tests focused on detecting signatures of balancing selection in genomic data
}
\keyword{internal}
