% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_alleles.R
\name{.count_alleles}
\alias{.count_alleles}
\title{Count derived or ancestral alleles for a given vcf position}
\usage{
.count_alleles(x, split = "|")
}
\arguments{
\item{x}{A vector containing 0s and 1s corresponding to allele states. 0/0: homozygous reference; 0/1: heterozygous; 1/1: homozygous alternate}

\item{split}{Pattern to look for. Inherited from stringr::str_split}
}
\description{
Count derived or ancestral alleles for a given vcf position
}
\examples{
data.table::data.table(col1="1|1", col2="0|0", col3="0|0") \%>\%
dplyr::summarise(across(col1:col3, .count_alleles))
}
