% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_ncd2.R
\name{.vcf_ncd2}
\alias{.vcf_ncd2}
\title{Make input file for NCD2 from VCF}
\usage{
.vcf_ncd2(x, nind = nind, index.col = index.col, verbose = T)
}
\arguments{
\item{x}{A data.table object generated from reading a VCF file with read_vcf.}

\item{nind}{A vector containing the number of diploid individuals from each
population to be used. The length of the vector should be equal to the number
of populations being sampled.}

\item{index.col}{First genotype column in VCF file.}

\item{verbose}{Logical. Printed updates. Default is T.}
}
\description{
Make input file for NCD2 from VCF
}
