% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncd1.R
\name{ncd1}
\alias{ncd1}
\title{Calculate non-central statistic (NCD1)}
\usage{
ncd1(
  x = x,
  tf = 0.5,
  fold = T,
  w = 3000,
  by.snp = T,
  ncores = 4,
  valormaf = "val",
  minIS = 10,
  label = "example_sim"
)
}
\arguments{
\item{x}{A data.table object}

\item{tf}{Target frequency}

\item{w}{Window size in bp.}

\item{by.snp}{Logical. If TRUE, windows are defined around each SNP
in the input data}
}
\value{
A data.table object
}
\description{
Calculate non-central statistic (NCD1)
}
