% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncd1.R
\name{ncd1}
\alias{ncd1}
\title{Calculate non-central statistic (NCD1)}
\usage{
ncd1(x = x, tf = 0.5, w = 3000, ncores = 2, minIS = 2, by = "POS")
}
\arguments{
\item{x}{A data.table object with
columns: CHR, POS, REF, ALT, tx_1 (number of alternate allele copies),
tn_1 (total number of alleles).}

\item{tf}{Target frequency.}

\item{w}{Window size in bp. Default is 3000.}

\item{ncores}{Number of cores. Increasing this can speed things up for you.
Default is 2.}

\item{minIS}{Minimum number of informative sites. Default is 2. Windows with
less informative sites than this threshold are discarded.}

\item{by}{Define how to scan the genome. "POS" (default) defined sliding windows based on w. Future implementation: "IS" defined windows around each informative site.}
}
\value{
A data.table object with columns:  Win.ID, S (sites), IS (informative sites), tf (target frequency), ncd1
}
\description{
Calculate non-central statistic (NCD1)
}
\examples{
ncd1(x=ncd1_input, tf=0.5, w=3000, ncores=2, minIS=2)
}
\references{
Bitarello et al. Genome Biology and Evolution, Volume 10, Issue 3, March 2018, Pages 939–955, https://doi.org/10.1093/gbe/evy054
}
