% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncd1.R
\name{ncd1}
\alias{ncd1}
\title{Calculate non-central statistic (NCD1)}
\usage{
ncd1(
  x = x,
  tf = 0.5,
  fold = T,
  w = 3000,
  by.snp = T,
  ncores = 4,
  valormaf = "val",
  minIS = 10,
  label = "example_sim"
)
}
\arguments{
\item{by.snp}{}
}
\description{
Calculate non-central statistic (NCD1)
}
