% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncd2.R
\name{ncd2}
\alias{ncd2}
\title{Calculate non-central statistic (NCD2)}
\usage{
ncd2(
  x = x,
  tf = 0.5,
  fold = T,
  w = 3000,
  by.snp = T,
  ncores = 4,
  valormaf = "val",
  minIS = 10
)
}
\arguments{
\item{by.snp}{}
}
\description{
Calculate non-central statistic (NCD2)
}
