% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncd2.R
\name{ncd2}
\alias{ncd2}
\title{Calculate non-central statistic (NCD2)}
\usage{
ncd2(x = x, tf = 0.5, w = 3000, ncores = 2, minIS = 2)
}
\arguments{
\item{x}{A data.table object}

\item{tf}{Target frequency}

\item{w}{Window size in bp. Default is 1000}

\item{ncores}{Number of cores. Increasing this can spead things up for you.
Default is 4.}

\item{minIS}{Minimum number of informative sites. Default is 2. Windows with
less informative sites than this threshold are discarded.}
}
\value{
A data.table object
}
\description{
Calculate non-central statistic (NCD2)
}
\examples{
ncd2(x=ncd2_input, tf=0.5, w=3000, ncores=2, minIS=8)
}
