% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vcf.R
\name{parse_vcf}
\alias{parse_vcf}
\title{Parse VCF files generated by SLiM v3}
\usage{
parse_vcf(
  infile = "*.vcf",
  outfile = NULL,
  nind = c(10, 10, 1),
  type = "ncd2",
  fold = T,
  intern = T,
  verbose = T
)
}
\arguments{
\item{infile}{The path and name of a vcf file - single chromosome}

\item{outfile}{The path and name for the outfile. If not provided,
this will be  a timestamp + infile in the current directory.}

\item{nind}{A vector containing the number of diploid individuals from each
population to be used. The length of the vector should be equal to the number
of populations being sampled.}

\item{type}{Which input format will be generated. Options: ncd1, ncd2.}

\item{fold}{Logical. If TRUE, the output will have alternate allele counts.
If FALSE, derived allele counts will be used. Default is TRUE.}

\item{intern}{Logical. If TRUE, output is kept in R in addition to being written
in output file. If FALSE, output is written to file only. Default is TRUE.}

\item{verbose}{Logical. If TRUE, progress reports will be printed as the
function runs.}
}
\value{
Returns a data table object.
}
\description{
Parse VCF files generated by SLiM v3
}
\examples{
parse_vcf(infile = "inst/example.vcf", nind = c(108, 1), fold=T)
}
