% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vcf.R
\name{parse_vcf}
\alias{parse_vcf}
\title{Parse VCF files generated by SLiM v3}
\usage{
parse_vcf(infile = NULL, n0 = NULL, n1 = NULL, type = NULL, verbose = F)
}
\arguments{
\item{infile}{The path and name of a vcf file - single chromosome}

\item{n0}{Number of individuals in pop0. Must be >=2 for both ncd1 and ncd2.}

\item{n1}{Number of individuals in pop1. Must be >=1 for ncd2. Must be NULL for ncd1. NULL by default.}

\item{type}{Which input format will be generated. Options: ncd1, ncd2.}

\item{verbose}{Logical. Printed updates. Default is T.}
}
\value{
Returns a data table object with columns CHR   POS REF ALT tx_1 tn_1 (if type is ncd1) or  CHR   POS REF ALT tx_1 tn_1 tx_2 tn_2 (if type is ncd2)
}
\description{
Parse VCF files generated by SLiM v3
}
\examples{
parse_vcf(infile=system.file(package="balselr", "example.vcf"), n0=108, type="ncd1")
parse_vcf(infile=system.file(package="balselr", "example.vcf"), n0=108, n1=1, type="ncd2")
}
