% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vcf.R
\name{parse_vcf}
\alias{parse_vcf}
\title{Parse vcf}
\usage{
parse_vcf(
  infile = NULL,
  vcf_data = NULL,
  n0 = NULL,
  n1 = NULL,
  type = NULL,
  verbose = F
)
}
\arguments{
\item{infile}{The path and name of a vcf file - single chromosome. Can be NULL if vcf_data is provided.}

\item{vcf_data}{A data.table object containing VCF data. Can be NULL if infile is provided.}

\item{n0}{Number of individuals in pop0. Must be >=2 for both ncd1 and ncd2.}

\item{n1}{Number of individuals in pop1. Must be >=1 for ncd2. Must be NULL for ncd1. NULL by default.}

\item{type}{Which input format will be generated. Options: ncd1, ncd2.}

\item{verbose}{Logical. Printed updates. Default is F.}
}
\value{
Returns a data table object with relevant columns based on type
}
\description{
Parse vcf
}
\examples{
parse_vcf(vcf_data=vcf_obj, n0=108, n1=2,type="ncd2")
parse_vcf(infile=system.file(package = "balselr", "example.vcf"), n0=108, type="ncd1")
}
