% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_ncd1.R
\name{vcf_ncd1}
\alias{vcf_ncd1}
\title{Make input file for NCD1 from VCF}
\usage{
vcf_ncd1(
  x = inp,
  outfile = outfile,
  nind = nind,
  index.col = index.col,
  verbose = verbose
)
}
\arguments{
\item{verbose}{}
}
\description{
Make input file for NCD1 from VCF
}
